# Ensure we're actually using a cached metadata provider in tests