--- a/test/metabase/query_processor/util/add_alias_info_test.clj
+++ b/test/metabase/query_processor/util/add_alias_info_test.clj
@@ -1032,9 +1032,8 @@
                                           (lib/with-join-fields :all))))))]
       (qp.store/with-metadata-provider mp
         (driver/with-driver :h2
-          (let [preprocessed (metabase.util/profile 'preprocess
-                               (-> question qp.preprocess/preprocess))
-                expected     (metabase.util/profile (add/add-alias-info preprocessed))] ; NOCOMMIT
+          (let [preprocessed (-> question qp.preprocess/preprocess)
+                expected     (add/add-alias-info preprocessed)]
             (testing ":source-query -> :source-query -> :joins"
               (is (=? [{:alias "Reviews"
                         :condition [:=