Originally appeared in #61398 

Lots of our tests run ridiculously slow because they're not **actually** using a metadata provider with a cache -- they're using a composed metadata provider that implements `CachedMetadataProvider` but without any of the underlying providers actually having caches.

BEFORE:

```
 ⮦ preprocess took 14.1 s {mb-test=metabase.query-processor.util.add-alias-info-test/model-duplicate-joins-test}
 ⮦ (add/add-alias-info preprocessed) took 22.2 ms {mb-test=metabase.query-processor.util.add-alias-info-test/model-duplicate-joins-test}
(model-duplicate-joins-test) took 14.2 s 
```

AFTER:

```
 ⮦ preprocess took 1.8 s {mb-test=metabase.query-processor.util.add-alias-info-test/model-duplicate-joins-test}
 ⮦ (add/add-alias-info preprocessed) took 14.2 ms {mb-test=metabase.query-processor.util.add-alias-info-test/model-duplicate-joins-test}
(metabase.query-processor.util.add-alias-info-test/model-duplicate-joins-test) took 1.9 s 
```

1.9 seconds for preprocessing is too slow IMO (FWIW this is in fact basically the hairiest query we have in our tests) but it's a lot better than 14.2 seconds (because we weren't caching anything and thus recalculating returned columns literally thousands of times). We can improve it more in the future, but at least future efforts to improve performance won't get tripped up by the `ensure-cached-metadata-provider` logic not working as intended.

This PR works by adding a new method to `CachedMetadataProvider`, `has-cache?`, that basically allows the `ComposedMetadataProvider` to decide whether or not it's an **actually-cached** metadata provider.

I know we could have avoided the new protocol method by just having two versions of `ComposedMetadataProvider`, a `CachedComposedMetadataProvider` if one of the underlying providers has a cache and an uncached version if it does not; but this seemed like a silly combinatorial explosion. Might change my mind later tho.