@camsaul 2025-08-05
Ok, so it only speeds it up a bit. Is that a reason not to approve the PR?